+++ packages/fuselage-ui-kit/src/Overflow.js
@@ -1,6 +1,5 @@
-import React, { useRef } from 'react';
+import React, { useRef, useCallback } from 'react';
 import {
-  AnimatedVisibility,
   Button,
   PositionAnimated,
   Options,
@@ -8,14 +7,26 @@ import {
   useCursor,
 } from '@rocket.chat/fuselage';
 
+import { useBlockContext } from './hooks';
+
 const convertOptions = (options, parser) => options.map(({ text, value }) => [value, parser(text)]);
-export const Overflow = ({ options, parser, onChange = console.log }) => {
-  const handleSelection = ([value]) => onChange({ target: { value } });
+
+export const Overflow = ({ context, options, parser, ...element }) => {
+  const [{ loading }, action] = useBlockContext(element, context);
+
+  const fireChange = ([value]) => action({ target: { value } });
   const convertedOptions = convertOptions(options, parser.text);
-  const [cursor, handleKeyDown, handleKeyUp, , [visible, hide, show]] = useCursor(-1, convertedOptions, (args, [, hide]) => {
-    handleSelection(args);
+  const [cursor, handleKeyDown, handleKeyUp, reset, [visible, hide, show]] = useCursor(-1, convertedOptions, (args, [, hide]) => {
+    fireChange(args);
+    reset();
     hide();
   });
+
+  const handleSelection = useCallback((...args) => {
+    fireChange(...args);
+    reset();
+    hide();
+  }, []);
   const ref = useRef();
   return (
     <>
@@ -26,12 +37,13 @@ export const Overflow = ({ options, parser, onChange = console.log }) => {
         onBlur={hide}
         onKeyUp={handleKeyUp}
         onKeyDown={handleKeyDown}
+        mod-loading={loading}
       >
         <Icon name='menu' size={20} />
       </Button>
       <PositionAnimated
         width='auto'
-        visible={visible ? AnimatedVisibility.VISIBLE : AnimatedVisibility.HIDDEN}
+        visible={visible}
         anchor={ref}
         placement='bottom right'
       >